Extract standalone js from compiled js
Gwt-compiled js is actually an html file that is intended to be loaded
within an iframe of a page (with some js in the main frame interacting
with it).

For DomDistiller, we want standalone javascript that we can inject into
the main frame.

This adds a python script and ant target that takes the compiled js and
converts it to standalone js for DomDistiller.

R=shashishekhar@chromium.org

Review URL: https://codereview.appspot.com/60350045