Add NodeListExpander and OrderedNodeMatcher
OrderedNodeMatcher takes a list of DOM nodes and provides a
match(Node n) that will match the nodes in order.

NodeListExpander provides a method to convert an ordered(*) list of
nodes into a minimal tree (from the DOM) containing those nodes-- i.e.
it will include all the ancestors of those nodes. This NodeTree has a
cloneSubtree method that will construct a copy of the corresponding part
of the DOM (by shallow-cloning each node).

This is required because Boilerpipe's output is essentially a list of
the text nodes that contain the content of the article. This will allow
us to extract the html elements for those nodes.

(*) In the order that they are encountered when walking the DOM tree.

R=nyquist@chromium.org

Review URL: https://codereview.appspot.com/61780043