Add html and image extraction to ContentExtractor
This makes DomToSaxParser calculate an in-order list of text nodes.
Boilerpipe's TextBlocks contain a list of the indexes of text nodes that
that block corresponds to. This allows us to map a TextBlock back to the
corresponding DOM nodes.

Then, this simply uses RelevantImageFinder and NodeListExpander
to get images and full html.

R=nyquist@chromium.org

Review URL: https://codereview.appspot.com/59220050