Fix crash when content extraction fails.
NodeListExpander.expand assumes that the list sent to it is non-empty.
That seems like a reasonable requirement for that function, so we should
only call it with a non-empty list. If we haven't identified any content
nodes, we can just return an empty string (node list expansion would
definitely not return anything better).

R=nyquist@chromium.org

Review URL: https://codereview.chromium.org/265393006