Change KeepLargestBlockFilter to be sibling-based instead of TagLevel-based.
TagLevel can be "broken" when a heading is merged with the subsequent content.
The TagLevel of the heading promotes all of the subsequent text to a lower tag
level (i.e. the tree rooted at the level of the heading). However, in many
cases, this causes our processing to exclude CONTENT level elements that are
immediately following other content (i.e. a series of <p> tags that didn't
get previously merged).

The new heuristic is essentially:
1) Compare the parent of the first text node to all prior text nodes to see if
they're siblings.
2) Compare the parent of the last text node to all subsequent text nodes to see
if they're siblings.

Elements identified in this manner are labeled SIBLING_OF_MAIN_CONTENT for tracking purposes.

BUG=378385

Sample urls:
http://articles.philly.com/2008-10-16/entertainment/24991990_1_calamari-squid-shark
http://www.livestrong.com/article/547660-how-to-absorb-zinc-better/
https://medium.com/design-startups/461772fbcac7
and several others.

R=cjhopman@chromium.org, nyquist@chromium.org

Review URL: https://codereview.chromium.org/319593005