Make DomDistiller's text-only output include all extracted text.
Because some elements are added to the output as part of RelevantElementsFinder,
they can bring in some text that's not in the document that Boilerpipe sees.
Changes DomDistiller to run the "text-only" output later, using the final HTML tree.
This also causes the text output to better match the distilled page in
that it now excludes the title.

R=cjhopman@chromium.org

Review URL: https://codereview.chromium.org/354723002