Don't merge non-content lists with trailing content.
BlockProximityFusion when *not* content-only, will take the last of a
set of non-content blocks and merge with subsequent content blocks.
While this is helpful in many cases for identifying a headline or
leading text, in the case of merging a non-content list item with
subsequent non-list text is almost never helpful. Examples are of the
form:

"""
optional useful text

<li>some ad
<li>more ad

More useful text
"""

Previously "more ad" would get included.

Addresses this specifically by limiting list merging. A subsequent CL
will merge lists more intelligently but this paves the way ensuring we
don't bring "some ad" in as well, if we just merged list items together.

Begins adding tests to BlockProximityFusion to help codify it's
behaviour.

BUG=378385
R=cjhopman@chromium.org, nyquist@chromium.org

Review URL: https://codereview.chromium.org/375873005