Instrument DomDistiller with timing information.
Introduces a TimingInfo message as part of the DomDistiller response
which includes timing breakdown of the various phases of processing as
well as total time.

Refactored ContentExtractor to break the phases down into smaller steps
and changed TextDocument to be the carrier of information accross phases
of processing.

R=cjhopman@chromium.org

Review URL: https://codereview.chromium.org/499623002