Add support for outputting debug info in result.
This CL adds support for outputting the debug log in the result proto.
It uses a generic proto message so it can be expanded later to include
other type of debug info.

It also cleans up some of the logging-related code by moving it to
LogUtil.

BUG=None
R=yfriedman@chromium.org

Review URL: https://codereview.chromium.org/497633002