don't prune a subtree when its only child is a text node.
bug:
http://homepages.tcp.co.uk/~maya/thug/thug7.html, from the cleaneval set,
has the entire page in 1 <pre>, with the DOM: <html> -> <body> -> <pre>.
NodeListExpander.pruneTopNodeChain() prunes away all these 3 subtrees,
leaving only the text node in <pre>.  Unfortunately, this is rejected by
ContentExtractor.formatExtractedNodes() because it's not an element node.

fix:
only prune subtrees whose only child is not a text node.
for the said url, the <pre> subtree will not be pruned, and hence will be
returned to and accepted by ContentExtractor.formatExtractedNodes().

BUG=408292
R=cjhopman@chromium.org

Review URL: https://codereview.chromium.org/521443002