Add SimilarSiblingContentExpansion
This filter identifies a "canonical" node for each TextBlock. This node
is the highest ancestor of the TextBlock's first contained node that
does not contain (in it's subtree) the last node of the previous
TextBlock or the first node of the next TextBlock. It then looks at
TextBlocks whose canonical node are siblings, and if those two
blocks/nodes are "similar enough"* and one of them is marked as content,
the other will be marked as content.

*"similar enough" has several toggles: is there a title between them? is
there a heading between them? how much link density does the non-content
one have? how far apart are they (in text blocks)? do the nodes have the
same tag?

This adds a single pass of such expansion (with dont-cross-titles,
dont-cross-headings, require-same-tag, max-ld-0.50, max-distance-3).

This then makes the first pass of BlockProximityFusion only fuse
content blocks (and simplifies the BlockProximityFusion logic a bit).

Here's some various scoring stats on reader+knowledge:
Master + the single pass of SimilarSiblingContentExpansion
https://x20web.corp.google.com/~cjhopman/domdistillerscore/score.2014-08-27-17.38.14.html

Master + the BlockProximityFusion change
https://x20web.corp.google.com/~cjhopman/domdistillerscore/score.2014-08-27-17.31.28.html

Master + this entire change
https://x20web.corp.google.com/~cjhopman/domdistillerscore/score.2014-08-27-17.29.35.html

(some unused files are deleted instead of updating them to compile in the new world)

R=nyquist@chromium.org, yfriedman@chromium.org

Review URL: https://codereview.chromium.org/515013002