Initial header image detection on articles.
Many articles have a large, article-width image above the main
content that is not detected by the distiller because it is
outside of the main content area and/or preceeds the article.

Several heuristics are used to rank and retrieve potential article
head images. Only images that have not already been marked as
content are considered. The following heuristics are used:
- DOM distance from first content node
- Image area (width*height)
- Image width/height ratio
- Has "figure" tag ancestor

Images are ranked in-place to retain their position in the DOM
and only a single image is chosen (assuming its score is above
a certain threshold). This addition is contained as part of
RelevantElementsFinder since the header image is exactly that; a
relevant element to the article.

BUG=427636
R=cjhopman@chromium.org

Review URL: https://codereview.chromium.org/688523003