Add content_image field to protobuff for scoring.
I have replaced a field in the DomDistillerResult protobuff called
"image_urls" that was almost entirely unused (one loop that never
ran inside of chromium). I have removed this field and replaced it with
a repeated "ContentImage" field that is designed to contain the
URL and, eventually, other information about the image. The primary
purpose of this change, currently, is for article scoring. Code has
also been added to populate this new field.

BUG=None
R=nyquist@chromium.org

Review URL: https://codereview.chromium.org/733573002