Make image source absolute.
This change is to fix a bug where image URLs are collected before
the distiller makes them absolute. The code is essentially the same
as before but has been moved to a new position. An additional small
test has been added to check this.

BUG=435371
R=nyquist@chromium.org

Review URL: https://codereview.chromium.org/739393006