Flesh out jstest environment
Adds SetUp/TearDown to JsTestCase and calls them as appropriate
in the test runner.

Adds a way to run a subset of the tests (though this isn't actually
hooked up to run_tests.py or `ant test` yet).

Adds reporting of # of tests, # failed, # skipped and total time (very
similar to the final line of the gwt tests output).

BUG=440102
R=mdjones@chromium.org

Review URL: https://codereview.chromium.org/807763002