Move ignoring elements into FilteringDomVisitor.
During TextDocument construction, we had multiple different ways to
ignore an element. Doing it in FilteringDomVisitor is best since it
causes us to skip the entire subtree (in the content handler we would
have to record our ignorable element depth and properly ignore some
events when in one).

This moves all the ignoring/skipping into the FilteringDomVisitor.

BUG=440977
R=nyquist@chromium.org

Review URL: https://codereview.chromium.org/792083004