Add StringUtil.countWords and simplify TextBlockBuilder
This makes TextBlockBuilder just track the number of words
inside/outside of anchors as it encounters text blocks (instead of
maintaining a stream of text + anchor start/end tokens and parsing
that at the end of the block).

Also uses the new countWords in a couple places where we were attempting
to count the words in a string.

BUG=440977
R=nyquist@chromium.org

Review URL: https://codereview.chromium.org/836573002