Change LogUtil.LOG_BUILDER to a String
A StringBuilder is implemented simply as an object with a javascript
string field. With a statically initialized StringBuilder a static
initializer method is created and called at the callsite of every
function in the class. Since LogUtil is used so broadly, this static
initializer is actually rather expensive (even though it doesn't do
anything after the first call).

Perf impact:
~3% decrease in total time.

BUG=440977
R=nyquist@chromium.org

Review URL: https://codereview.chromium.org/791053008