Move PagingLinksFinderTest from test/ to jstest/
The "no number" rule is messed up in test/ environment because the IP
and port are considered numbers. Tests test1NextLink and
test1PrevAnd1NextLinks are updated accordingly.

This is the last test case in test/, so TestUtil.java in test/ can be
removed as well.

BUG=440102
R=cjhopman@chromium.org

Review URL: https://codereview.chromium.org/852683002