Add a WebDocument class
The WebDocument is a simplified view of the web page (currently only contains
WebText elements, but will be extended with other things we are interested
in like images, tables, embeds, etc).

Basically, the process of extracting content has three distinct parts for each
type of content that we want to extract:
1. identify the logical elements (text, image, video, etc.)
2. determine what is content
3. extract a represententation of the content elements

The WebDocument moves us to a world where we do those three steps somewhat
consistently for all different types of elements. First, we create the
WebDocument which requires doing 1 for everything. Then do processing on that
to detect content. Then basically just iterate through the content elements to
extract their representation.

R=nyquist@chromium.org

Review URL: https://codereview.chromium.org/838493003