Added README.md pointing to Google Code wiki.
As a placeholder, this CL adds a README.md file that points to the
Google Code wiki. The content of the wiki will later be moved to
this README.md file.

This also serves as a test for committing changes using Rietveld,
'git cl land' and GitHub.

R=cjhopman@chromium.org

Review URL: https://codereview.chromium.org/883423002