Use numbers as a fast filter in next page detection
Existence of numbers is done after matching baseUrl against the link
URL, which is accurate but slower. If a link has no number at all in
its path part, it won't match for next page, and is fast to filter
out.

After this optimization, pagination time is reduced to 75%. Combined
with the URL prefix filter, the pagination time is reduced to around
60% * 75% = 45%.

BUG=431067
R=cjhopman@chromium.org

Review URL: https://codereview.chromium.org/899093004