Don't let base tag change original_url
When base tag wrongly changes original_url, wndLocationHref would
be wrong, and if the link href collide with wndLocationHref, that
link would be wrongly disqualified. Practically this bug is unlikely
to change output because base tag often points to the root path of
the host, which is unlikely to be the next page.

This fixes ff4bc3d "Fix relative link resolution with base tag".

R=cjhopman@chromium.org

Review URL: https://codereview.chromium.org/925893002