Fix flakiness due to a bug in createAnchorWithBase()
The anchor element was not inserted into the new document with the
base tag. However, the result was correct most of the time, which
is quite surprising.

This bug was introduced in revision 7b12bc7 "Resolve links in a new
HTML document with a base tag".

R=cjhopman@chromium.org

Review URL: https://codereview.chromium.org/955193002