Show distilled page in Chrome Extension
After clicking the Reader Mode icon in the omnibox, the distilled page
would replace the original one, just for debugging purpose.
Clicking "Profile Extraction" button in the devtool panel would not
change the original page, so it is still possible to do profiling for
multiple times.

The outer console.log() is not very useful, so removed.

R=cjhopman@chromium.org

Review URL: https://codereview.chromium.org/1133543002