BR elements should be folded into WebText
WebText will clone its parent tag to maintain the structure of the
document when generating output. This means that if the BR element
flushes the current WebText, extra copies of the parent tag will be
added for each BR tag. For instance:

<p>item1<br>item2<br>item3</p>

will result in

<p>item1</p><br><p>item2</p><br><p>item3</p>

Instead of splitting WebText when we see BR elements, we should fold
them into the contents of the WebText. This prevents BR elements from
surrounding lines with copies of the parent tag.

BUG=485829
R=mdjones@chromium.org, wychen@chromium.org

Review URL: https://codereview.chromium.org/1143633002

Patch from mrobinson <mrobinson@igalia.com>.