Avoid chromedriver elsewhere if any
Make sure the chromedriver installed by install-build-deps.sh is used.
Before this change, if chromedriver is already installed and is in $PATH,
that would be used instead of the one installed by install-build-deps.sh.

Also downgrade chromedriver to version 2.8, in order to be compatible
with the Chrome on drone.io (M30 dev) and the current one (M45) at the
same time. chromedriver doesn't have to be pinned to 2.8 after "Update
Chrome if it is too old" (https://crrev.com/1159743006/).

R=kuan@chromium.org, mdjones@chromium.org

Review URL: https://codereview.chromium.org/1148083004