Update Chrome if it is too old
The default Chrome channel is kept intact, to avoid possible conflict
during the update.

The accepted minimal version of Chrome is M32, when URL() is supported.
Ref: https://developer.mozilla.org/en-US/docs/Web/API/URL

Sandbox is disabled to avoid crashing in continuous integration.
Ref: "Allow tests to run without Chrome sandbox"
(https://crrev.com/1164683006/)

This depends on "Avoid chromedriver elsewhere if any"
(https://crrev.com/1148083004/) because the installed version of
chromedriver on drone.io is 2.1, which is too old for current version of
Chrome.

BUG=483149
R=nyquist@chromium.org

Review URL: https://codereview.chromium.org/1159743006