Leverage semantic elements to find the main article
Semantic web elements or metadata can tell us where the main article
is in a structured way. Only processing that element is a fast path
to boost the speed of distillation.

In a well-formed html5 page of an article, the <article> element can
directly be used for distillation. The processing is much faster this
way.

Similarly, schema.org metadata can give a strong hint where the main
article is. This is often less precise than the <article> element,
so this signal is used in a different way.

** Score changes:
- Content:
    reader-mode-golden-data:
      Average precision +0.003, recall -0.001.
- Images:
    reader-images-golden-data:
      Average precision +0.012, recall +0.005.
- Next page:
    page-links-golden-data:
      No change.

** Performance improvement:
- reader-mode-golden-data:
    Spent 15% less time on average.

BUG=431067
R=cjhopman@chromium.org

Review URL: https://codereview.chromium.org/1131793009.