Fix for keeping lists structure
Created the WebTag class to represent the tags we want to preserve over
the distillation process. A WebTag instance acts like a placeholder
which is injected to the WebDocument when walking the DOM and is used to
track the UL, OL and LI element positions.

This mechanism could be used in the future for all the tags that we want
to preserve.

When the generateOuput() is called for each WebTag, it returns the html
tag that the instance is representing.  A new filter was created to
process these WebTags and mark as content considering its content
inside. This filter runs for the last since it relies on content found
by all the others filters.

BUG=502524
R=mdjones@chromium.org, wychen@chromium.org

Review URL: https://codereview.chromium.org/1230583006 .

Patch from marcelorcorrea <marcelorcorrea@gmail.com>.