Remove //base dependency
The C header file for proto JSON was generated here, checked in to the
distribution repository, and rolled into Chromium. It is not allowed to
have //base dependency on external repos, so the file generation should
be moved to Chromium.

The ant target "protoc.json" is removed, and plugin_pb2.py is exported
to the distribution repo.

This goes along with https://codereview.chromium.org/2034373002/

BUG=572585
R=mdjones@chromium.org, nyquist@chromium.org

Review URL: https://codereview.chromium.org/2039043003 .