Make sure images are loaded before running tests
The data URI images began to be decoded asynchronously since
https://crbug.com/514206 was fixed. If tests are run before images are
decoded, the dimension would be wrong.

There was a racing condition between starting the test and decoding the
image, and the tests were flaky. The failure rate could be as high as
10% recently.

Reload the HTML every time before running the tests and make sure they
are loaded before running the tests. Due to timing issues, image_loaded
always returned True in 4000 local runs.

This fixes flakiness in https://crrev.com/2179593002 (commit 6c16f14).

BUG=
R=mdjones@chromium.org

Review-Url: https://codereview.chromium.org/2755503005 .