Update patch set 1
Patch Set 1:

I am working on enabling --use-zoom-for-dsf for Android, which merges the device scale factor into the page scale factor (For details, please see crbug.com/485650 and https://docs.google.com/document/d/1Yl6ZsDn36jImtu4JnAawxvqRsl2aWa3OhociZHw7iT8/edit).
Now I am trying to enable it as default on Android. However, I met a unit test failure caused by DomDistillerJsTest.RunJsTests because of a precision issue.

In EmbedExtractorTest.testImageExtractorWithAttributesCSSHeightCM and testImageExtractorWithAttributesCSSHeightCMAndWidthAttrb, the height is expected to be 38, but it is 37 when the device scale factor is 2.625. WebImage.getHeight() returns a value that is from HTMLImageElement::LayoutBoxHeight() in Blink layer, which returns floor(99.501 / 2.625) == floor(37.905142857142856) == 37. While HTMLImageElement::LayoutBoxHeight() conducts the floor against the height, Element::clientHeight() rounds it and gets the more accurate value.
This CL replaces WebImage.getHeight() used in EmbedExtractorTest.testImageExtractorWithAttributesCSSHeightCM and testImageExtractorWithAttributesCSSHeightCMAndWidthAttrb to WebImage.getClientHeight() which return the same value of Element::clientHeight().

Please review this code.

Patch-set: 1