Allow small error for image height test
EmbedExtractorTest.testImageExtractorWithAttributesCSSHeightCM fails when
enabling --use-zoom-for-dsf on Android because of a precision issue. This CL
lets EmbedExtractorTest.testImageExtractorWithAttributesCSSHeightCM allow small
error against the expected value of the height.

TEST=components_browsertests DomDistillerJsTest.RunJsTests on Android and Linux
Chromes using NEXUS 5X and desktop. Dom-distiller itself (using its test cases)
is also tested on Linux. On desktop, the test runs with the device scale factor
of 1.75, 2, and 2.625 (i.e., --force-device-scale-factor=DSF).

Bug: 737777
Change-Id: I634fafe54594e79993da61e347eda022f865ce31